fixed aggregator typing isue with voer saple array

+++ autogalaxy/imaging/model/plotter_interface.py
@@ -117,8 +117,8 @@ class PlotterInterfaceImaging(PlotterInterface):
                 dataset.data.native_for_fits,
                 dataset.noise_map.native_for_fits,
                 dataset.psf.native_for_fits,
-                dataset.grids.lp.over_sample_size.native_for_fits,
-                dataset.grids.pixelization.over_sample_size.native_for_fits,
+                dataset.grids.lp.over_sample_size.native_for_fits.astype("float"),
+                dataset.grids.pixelization.over_sample_size.native_for_fits.astype("float"),
             ]
 
             hdu_list = hdu_list_for_output_from(
